% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mcmm.glmnet.R
\name{mcmm.glmnet}
\alias{mcmm.glmnet}
\title{Multiple Core Multiple Model glmnet models fitting}
\usage{
mcmm.glmnet(FUN, ...)
}
\arguments{
\item{FUN}{a function of appropriate form to generate data to fit the model}

\item{...}{a parameters to be passed in function}

\item{family}{a parameter that controls the type of model to be fitted.}
}
\value{
Fitted glmnet models for generated data
}
\description{
Performs fitting simple L1 penalized models of choosen type foor generated data.
Data set with n observations and p predictors is constructed.
Models are fitted for each p-1 predictors using multi core lapply function.
}

